requirements.txt
openai
tqdm
ffmpeg-python
pillow
